+++ apps/aurora-portal/src/server/Network/routers/index.ts
@@ -1,6 +1,5 @@
 import { floatingIpRouter } from "./floatingIpRouter"
 import { networkRouter } from "./networkRouter"
-import { portRouter } from "./portRouter"
 import { securityGroupRouter } from "./securityGroupRouter"
 import { securityGroupRuleRouter } from "./securityGroupRuleRouter"
 import { rbacPolicyRouter } from "./rbacPolicyRouter"
@@ -10,7 +9,6 @@ export const networkRouters = {
   network: auroraRouter({
     ...networkRouter,
     floatingIp: floatingIpRouter,
-    port: portRouter,
     securityGroup: securityGroupRouter,
     securityGroupRule: securityGroupRuleRouter,
     rbacPolicy: rbacPolicyRouter,
